requirements.txt
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk
python-dotenv
requests
